+++ src/app/pages/Savings/SavingsPage.tsx
@@ -2,16 +2,11 @@ import React from 'react';
 import { useRouteMatch } from 'react-router';
 import { Link } from 'react-router-dom';
 
-import { useApi } from 'services/api';
-import { useSubscribable } from 'utils/react';
 import { routes } from 'app/routes';
 import { makeStyles } from 'utils/styles';
-import { TabsList, TabContext, Tab, TabPanel, Loading, Grid } from 'components';
-import { SwitchInput } from 'components/inputs';
-import { liquidityAmount } from 'utils/mock';
+import { TabsList, TabContext, Tab, TabPanel, Grid } from 'components';
 
-import { Card, WithViewDetails } from './Card/Card';
-import { DepositToPoolForm } from '../Demo/components/DepositToPoolForm';
+import { AllocateTab } from './AllocateTab';
 
 export function SavingsPage() {
   const match = useRouteMatch<{ page: string }>('/savings/:page');
@@ -32,8 +27,6 @@ export function SavingsPage() {
   return renderTabs();
 
   function renderTabs() {
-    const api = useApi();
-    const [pools, poolsMeta] = useSubscribable(() => api.savings.getPools$(), [api]);
     return (
       <Grid className={classes.root}>
         <TabContext value={selectedPage}>
@@ -56,42 +49,7 @@ export function SavingsPage() {
             </TabsList>
           </div>
           <TabPanel value={routes.savings.allocate.getElementKey()}>
-            <Loading meta={poolsMeta}>
-              <Grid
-                container
-                direction="row"
-                justify="flex-start"
-                alignItems="flex-start"
-                spacing={3}
-              >
-                {pools &&
-                  pools.map(pool => (
-                    <Grid item>
-                      <Card
-                        title="Compound"
-                        balanceAmount={liquidityAmount}
-                        liquidityAmount={liquidityAmount}
-                        tokens={pool.tokens}
-                        footerElement={
-                          <WithViewDetails
-                            getLink={() => '123'}
-                            allocateSwitcher={<span><SwitchInput />Allocate</span>}
-                            additionalElement={
-                              <DepositToPoolForm poolAddress={pool.address} supportedTokens={pool.tokens} />
-                            }
-                          />
-                        }
-                      />
-                    </Grid>
-                  ))}
-                <Grid item>
-                  <Card title="Y Curve" balanceAmount={liquidityAmount} liquidityAmount={liquidityAmount} tokens={[]} />
-                </Grid>
-                <Grid item>
-                  <Card title="Compound" balanceAmount={liquidityAmount} liquidityAmount={liquidityAmount} tokens={[]} />
-                </Grid>
-              </Grid>
-            </Loading>
+            <AllocateTab />
           </TabPanel>
           <TabPanel value={routes.savings.withdraw.getElementKey()}>
             Withdraw not implemented
@@ -117,16 +75,6 @@ const useStyles = makeStyles(
       display: 'flex',
       alignItems: 'flex-start',
       justifyContent: 'space-between',
-    },
-    cardsContainer: {
-      paddingTop: 60,
-      display: 'flex',
-      justifyContent: 'space-between',
-      alignItems: 'start',
-      flexWrap: 'wrap',
-    },
-    cardWrapper: {
-      margin: '0px 20px 40px 0',
     },
   }),
   { name: 'SavingsPage' },
